Store terraform state in s3
Since we now wish to store terraform state in s3, we've created a bootstrap
script to create an s3 bucket to store that state. We also create a
DynamoDB table to lock deployments to prevent simultaneious changes.

The script requires three variable to be set defining the bucket name,
DynamoDB table name and AWS region for those resources. Running the script
with these set will create the resources.

Following this, the local terraform state is migrated to the new backend
using `make terraform-init`. The first time this is run there's a prompt
to copy existing state to s3 - we do!

It is now safe to delete the local .tfstate files.