Promote changes to prod through staging
CI is changed such that pushes to main will trigger the staging
pipeline. If the staging pipeline is successful, the branch "staging"
is updated to the current SHA. The prod pipeline will trigger on pushes
to the staging pipeline.

Environments are now configured via the environments.yml file. Most
operations will fail when targeting environments not referenced in this
file.